pyproject.toml
[tool.poetry]
name = "python-matlab-ray-example"
version = "0.1.0"
description = "An example python project that calls matlab function in remote AWS instance"
authors = ["Sota Sato <sotasato@nii.ac.jp>"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "~3.8"
ray = {extras = ["default"], version = "^2.3.1"}
boto3 = "^1.26.116"
matlabengine = {version = "^9.14.2", optional = true}
numpy = "^1.24.2"
pandas = "^2.0.0"

[tool.poetry.group.dev.dependencies]
jupyterlab = "^3.6.3"
mypy = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

